+++ src/misc/networking/tcp-server.hpp
@@ -56,7 +56,7 @@ class TcpConnection
     std::atomic<bool> disconnectNotified{false};
     TcpClientInfoHandle clientInfoHandle;
     RcvCmdState rcvCmdState = RcvCmdState::ParseCmd0;
-    uint16_t rcvCmdLen = 0;
+    uint32_t rcvCmdLen = 0;
     uint16_t lastDataStart = 0;
     net::CmdQueueData rcvdCmd;
 };
